Feat/ble connection with phone

+++ src/ml_sensor/ml_sensor.h
@@ -1,14 +1,13 @@
 #ifndef ML_SENSOR_HEADER
 #define ML_SENSOR_HEADER
 
-#include<zephyr/drivers/adc.h>
-#include "../util/util.h"
+#include "util.h"
 
 
 static struct ml_internal_struct
 {
     bool convertion_done;
-    float uv_index;
+    uint8_t uv_index;
 };
 
 void ml_worker(void);
